requirements.txt
streamlit==1.22.0
opencv-python==4.9.0.80
Pillow==10.0.0
numpy==1.24.3
tensorflow==2.13.0
scikit-learn==1.3.0
